Team1 to DEV

+++ app/src/main/java/studentcompany/sportgest/domains/Game.java
@@ -6,14 +6,14 @@ public class Game extends DomainPojo {
     private long id;
     private Team home_team;
     private Team visitor_team;
-    private int date;
+    private long date;
     private String report;
     private int home_score;
     private int visitor_score;
     private float duration;
 
 
-    public Game(long id, Team home_team, Team visitor_team, int date, String report, Integer home_score, Integer visitor_score, float duration) {
+    public Game(long id, Team home_team, Team visitor_team, long date, String report, Integer home_score, Integer visitor_score, float duration) {
         this.id = id;
         this.home_team = home_team;
         this.visitor_team = visitor_team;
@@ -24,7 +24,7 @@ public class Game extends DomainPojo {
         this.duration = duration;
     }
 
-    public Game( Team home_team, Team visitor_team, int date, String report, Integer home_score, Integer visitor_score, float duration) {
+    public Game( Team home_team, Team visitor_team, long date, String report, Integer home_score, Integer visitor_score, float duration) {
         this.home_team = home_team;
         this.visitor_team = visitor_team;
         this.date = date;
@@ -59,7 +59,7 @@ public class Game extends DomainPojo {
         this.visitor_team = visitor_team;
     }
 
-    public int getDate() {
+    public long getDate() {
         return date;
     }
 
@@ -122,7 +122,7 @@ public class Game extends DomainPojo {
         int result = (int) (id ^ (id >>> 32));
         result = 31 * result + (home_team != null ? home_team.hashCode() : 0);
         result = 31 * result + (visitor_team != null ? visitor_team.hashCode() : 0);
-        result = 31 * result + date;
+        result = (int) (31 * result + date);
         result = 31 * result + (report != null ? report.hashCode() : 0);
         result = 31 * result + home_score;
         result = 31 * result + visitor_score;
